Initial MinSeg motor control/odometry tracking
Serial commands can now be sent to the MinSeg bot to set the direction
of movement (stop, forward, reverse, turn left, turn right). The
displacement of the robot is now tracked using the NXT motor encoders
and can be read through serial communications. Accuracy of the
displacement readings has not been completely tested. Further
description of the process used to read the encoders can be found in
comments of minsegmega.ino.